+++ src/paralleltasks/PopulateGridTask.java
@@ -18,10 +18,10 @@ public class PopulateGridTask extends RecursiveTask<int[][]> {
     CensusGroup[] censusGroups;
     int lo, hi, numRows, numColumns;
     MapCorners corners;
-    float cellWidth, cellHeight;
+    double cellWidth, cellHeight;
     final static int SEQUENTIAL_CUTOFF = 10000;
 
-    public PopulateGridTask(CensusGroup[] censusGroups, int lo, int hi, int numRows, int numColumns, MapCorners corners, float cellWidth, float cellHeight) {
+    public PopulateGridTask(CensusGroup[] censusGroups, int lo, int hi, int numRows, int numColumns, MapCorners corners, double cellWidth, double cellHeight) {
         throw new NotYetImplementedException();
     }
